View refactoring

+++ src/main/java/ru/olegcherednik/zip4jvm/io/in/DataInput.java
@@ -44,7 +44,7 @@ public interface DataInput extends Closeable {
 
     byte[] readBytes(int total) throws IOException;
 
-    void skip(int bytes) throws IOException;
+    void skip(long bytes) throws IOException;
 
     long length() throws IOException;
